Change Animations SearchView, create send data to Detailes Fragment, code have to refactor

+++ app/src/main/java/com/example/krakowautobusy/api/Api.kt
@@ -90,7 +90,7 @@ fun getInfoAboutLinePatternAnyVehicleStop(patternName:String):ArrayList<LineData
 
     fun getInfoAboutLineConcretDirectionLastStopName(
 
-        numberLine: Long,
+        numberLine: Int,
         lastStopName: String
     ): LineData {
