+++ app/src/main/java/com/example/krakowautobusy/database/LineInteerface.kt
@@ -3,7 +3,7 @@ package com.example.krakowautobusy.database
 import android.database.sqlite.SQLiteDatabase
 
 interface LineInteerface {
-    fun getInfoAboutLineConcretDirectionName(db: SQLiteDatabase, numberLine:Long, lastStopName:String):LineData
+    fun getInfoAboutLineConcretDirectionName(db: SQLiteDatabase, numberLine: Int, lastStopName:String):LineData
     fun getInfoAboutLineConcretDirectionId(db: SQLiteDatabase, numberLine:Long, lastStopId:Long):LineData
     fun getInfoAboutLinesAnyDirection(db: SQLiteDatabase, numberLine:Long):ArrayList<LineData>
     fun getInfoAllLine(db:SQLiteDatabase):ArrayList<LineData>
